refactor: move and generalize SearchFilterBar
