+++ src/components/admin/AdminCourseCatalog.vue
@@ -4,7 +4,7 @@ import { defineAsyncComponent } from 'vue'
 import { X, Book, Plus } from 'lucide-vue-next'
 import AdminCourseAddModal from '@/components/modals/AdminCourseAddModal.vue'
 import AdminCourseEditModal from '@/components/modals/AdminCourseEditModal.vue'
-import AdminSearchFilterBar from '@/components/admin/AdminSearchFilterBar.vue'
+import AdminSearchFilterBar from '@/components/SearchFilterBar.vue'
 import AdminPagination from '@/components/admin/AdminPagination.vue'
 import { useCoursesStore } from '@/stores/coursesStore'
 import type { Course, CourseInstructor, Person } from '@/interfaces/interfaces'
